fix(fst-dict): remove duplication by importing types from parsing crate
Types include:
- Span
- CharString
- WordMetadata
