+++ harper-core/src/document.rs
@@ -7,9 +7,9 @@ use paste::paste;
 use crate::parsers::{Markdown, Parser, PlainEnglish};
 use crate::patterns::{PatternExt, RepeatingPattern, SequencePattern};
 use crate::punctuation::Punctuation;
-use crate::span::Span;
 use crate::token::NumberSuffix;
 use crate::vec_ext::VecExt;
+use crate::Span;
 use crate::{Dictionary, FatToken, FstDictionary, Lrc, Token, TokenKind, TokenStringExt};
 
 /// A document containing some amount of lexed and parsed English text.
